runtime_vm: Avoid premature container info cleanup on creation failure
While the removing the container info of the failed container from the
containers' map seems harmless, it'll end up causing an error when
actually cleaning up / deleting the failed container.

The reason for this is that CreateContainer, on Kata side, will remove
the failed container, wich will result on DeleteContainer, on the CRI-O
side, being called and then failing to actually remove the container, as
its info has already been removed from the containers' map.

This issue has been reported as part of a PR opened by Evan Foster:
https://github.com/kata-containers/runtime/pull/2826#issuecomment-657754302

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Tested-by: Evan Foster <efoster@adobe.com>