node: fix panic if /sys/fs/cgroup failed to stat
our rhel 7 boxes that are used to package cri-o are failing with the following error:
```
panic: cannot statfs cgroup root
goroutine 1 [running]:
```
this is because the libcontainer cgroup packages opts to panic when the stat on /sys/fs/cgroup fails, rather than return an error

The point of the node package is to gracefully catch and handle these errors, so we switch to libpod version instead, and will now print
we failed to stat, rather than panic on it.

Note: these nodes are not suitable for running workloads, as runc will later panic, but packaging (in this case, building a config) should not panic because of it.

Signed-off-by: Peter Hunt <pehunt@redhat.com>