Update the container resources under the spec
Call to the `UpdateContainerResources` did not update the resources under the container spec,
so we had a situation when resources had updated values under the cgroup, but it did not reflect
under the container spec.

Now `UpdateContainerResources` will update the container spec, in the case when the update will fail,
the code will rollback old spec values.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>