runtime_vm: Cleanup process when the Container is Stopped
When the container finishes its execution a containerd-shim-kata-v2
process is left behind (only when using CRI-O).

The reason for that seems to be CRI-O not doing a cleanup of the process
whenever the container state has changed its state from running to
stopped.

The most reasonable way found to perform such cleanup seems to be taking
advantage of the goroutine used to update the container status and
performing the cleanup there, whenever it's needed.

Related: https://github.com/kata-containers/runtime/issues/2719

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>