Restore Sandbox createdAt timestamp on cri-o restart
When restarting cri-o, try to restore the createdAt timestamp of
Sandboxes from the annotations. Previously the timestamp was reset to
time.Now() when restoring Sandboxes. As Sandboxes are not guaranteed to
be restored in creation order in server.restore() this could lead to
ordering issues in calls like e.g. ListPodSandbox which would not return
the Pods in the orginial creations order.

Signed-off-by: Ralf Haferkamp <rhafer@suse.com>