Add `make release` target
The new `make release` target can be used to automatically create a PR
which bumps the CRI-O version. The fork organization has to be set
manually like this:

```
> make release ORG=saschagrunert
```

We are also able to specify the remote name for which we wanna push the
changes:

```
> make release ORG=saschagrunert REMOTE=other-than-origin
```

The script will automatically pick-up the local branch (the target
branch for the PR), increment the version in version.go and create the
PR for that.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>