test/*bats: fix shellcheck SC2076
This fixes shellcheck warning SC2076, for example

> In default_mounts.bats line 90:
>     [[ "$output" =~ "test.txt" ]]
>                     ^--------^ SC2076: Don't quote right-hand side of =~, it'll match literally rather than as a regex.

So, operator ~= does a comparison against a regular expression, and yet
in all the places we are actually using it to search for a substring.

Use the substring syntax (i.e. [[ "$output" == *"substring"* ]]) instead
to be more clear about what we're doing.

This commit is courtesy of

for f in *.bats; do sed -i -e 's/ =\~ "\(.*\)" ]]/ == *"\1"* ]]/g' $f; done

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>