Revert "tests: adjust test to not depend on runc behavior"
First, the code

	[ "$status" -ne 0 ] || [ "$output" =~ "Exit code: 1" ]

is not working since one can't use =~ operator inside [..]
(it is only working inside [[..]]).

Here's an example:

	$ [ "aaa bbb" =~ "aaa" ]; echo $?
	bash: [: =~: binary operator expected
	2

Since the test never failed, I guess $status was always non-zero,
so the second part was never evaluated, thus it is not needed.

This was found by shellcheck:

```
In ctr.bats line 1008:
	[ "$status" -ne 0 ] || [ "$output" =~ "Exit code: 1" ]
                                           ^-- SC2074: Can't use =~ in [ ]. Use [[..]] instead.
```

Second, the cleanup/stop stuff added by the commit is not needed,
since teardown() is executed after each test, and it contains
a call to cleanup_test() which calls the same 3 functions that
were added.

This reverts commit 2649ebb1a5d39208f8d2d68303e42a81048f4bdf.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>