test/*bats: fix excessive use of export
The `export VAR=value` is needed when used from functions like
`setup()`, since bats runs every @test case in a subshell.

In other places (such as when run from @test case) this is not needed,
as we don't have subshells to pass the VAR to.

More to say, in all the cases in this commit VAR is only needed by
`start_crio`, so instead of setting it for the whole test, set it
for the `start_crio` invocation only.

This fixes a bunch of shellcheck warnings SC2030,SC2031.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>