server: reduce complexity of getSourceMount
The function, as it was written, iterates over mountinfo slice multiple
times. In fact, a single pass is sufficient -- we just need to find the
mounts which can be a parent of the argument, and select the longest
one.

A test case is added to ensure correctness.

This should speed up container creation a notch.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>