Fix container cgroup under cgroupfs
In case cgroupfs cgroup manager is used, the container got created under
the conmon cgroup (rather than under the root), and then the stats
doesn't work because the cgroup path is invalid.

For example, if the conmon cgroup is created as
/sys/fs/cgroup/pod_123-456/crio-conmon-<ID>, then the container cgroup
is created as
/sys/fs/cgroup/pod_123-456/crio-conmon-<ID>/pod_123-456/crio-<ID>,
rather than sys/fs/cgroup/pod_123-456/crio-conmon-<ID>.

The `cricrl stats` then fails as it can't find the container cgroup.

This happens because the cgroupsPath set in spec for such containers
is not absolute, so the cgroup is created under the current cgroup,
which is that of conmon.

The fix is easy.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>