Use Unmount w/o prior Mounted check
The mount.Mounted check is super expensive, as it has to parse the whole
nine yards of /proc/self/mountinfo, which is slow and sometimes racy
(for latter, see https://github.com/kolyshkin/procfs-test).

It is much simpler and faster to just try unmount, and ignore EINVAL
error. Note that EINVAL might be returned if the argument is not a mount
point, or if flags are wrong. Since the flags is hardcoded here, the
code will only get EINVAL in case the path is not mounted.

This should speed things up a little, and fix an occasional case of real
mount not being unmounted because of the mountinfo race.

Also, since we do not check the path in mountinfo, and umount(2) deals
just fine with symlinks, we no longer need to resolve the path, so
remove the path cleaning code (which in fact should have been
filepath.EvalSymlinks()).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>