test/cgroups.bats: enable pids test for cgroup v2
The way to check if pids cgroup controller is available was only working
for cgroupv1, so the test was errorneously skipped on v2.
Fix to use a check that works for both v1 and v2.

Fix the "get the limit from inside the container" command, too, to work
for v1 and v2.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>