Add fidencio to OWNERS file
I've been working in the VM runtime type, and would like to officially
become responsible for this area.

Contributions made so far:
* runtime_vm: Apply the correct label before the sandbox is created
  - https://github.com/cri-o/cri-o/pull/3889
* Add runtume_type as an option of "--runtimes"
  - https://github.com/cri-o/cri-o/pull/3903
* runtime_vm: Ignore ttrpc.ErrClosed when shutting the container down
  - https://github.com/cri-o/cri-o/pull/3924
* runtimeVM: Improve CreateContainer cleanup in case of failures
  - https://github.com/cri-o/cri-o/pull/3956
* runtime_vm: Avoid possible deadlock on UpdateContainerStatus()
  - https://github.com/cri-o/cri-o/pull/3987
* runtimeVM: Cleanup a "Completed" container
  - https://github.com/cri-o/cri-o/pull/3998
* runtimeVM: Store logs again, and store them in the correct format
  - https://github.com/cri-o/cri-o/pull/4082
* And a bunch of backports ...

My main goal is to have the VM runtime type in a good shape so we can
have it properly tested & used on the kata side.

Together with the addition to the OWNERS file, I'm also adding myself to
the CODEOWNERS file, in order to automatically be added as a reviewer to
pull requests related to the VM runtime type.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>