managed ns: ignore `PID not initialized` on sandbox creation
everytime we create a sandbox with a pinned namespace, we spit out an error that says:
`Pid for infra container $id not found: PID not initialized`

in the pinned namespace case, we create the namespaces, create the networking stack (giving the created net namespace to the plugin), THEN create the pod.
Thus, we don't expect the infra container to be initialized.

Fix this by ignoring ErrNotInitialized when we're fetching the infraPid when finding the namespaces.

We can safely do this because the managed namespace case is covered (we don't use the pid, but use the managed namespace struct we've created)
And in the non-managed namespace case, we will (correctly) fail later because we return an invalid PID, which matches our invalid infra container.

Signed-off-by: Peter Hunt <pehunt@redhat.com>