test/network.bats: simplify, fix shellcheck
1. Simplify the code.

For example,

	run crictl start ...
	echo "$output"
	[ "$status" -eq 0 ]

is equivalent to

	crictl start ...

as it much easier to read (yes, bats runs tests with set -e
so there is an error check in place).

One more example,
	run crictl runp ...
	echo "$output"
	[ "$status" -eq 0 ]
	pod_id="$output"

is roughly equivalent to

	pod_id=$(crictl runp ...)

(except it does not print stdout, only stderr).

Another example,

	cfg=$(python .... < $input)
	echo "$cfg" > $output

is equivalent to

	python ... < $input > $output

and avoids the use of a temporary variable (saves some memory).

Finally,

	run crictl start ...
	echo "$output"
	[ "$status" -ne 0 ]

is equivalent to

	crictl start ... && false

2. Optimize use of CONMON_BINARY variable, and remove unneded cleanup.

3. Fix indentation, add some vertical whitespace for better readability.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>