config: set internal RootConfig to default storage if not specified
After we pull in the initial config, it is our source of truth for telling the world where our container storage is
in particular, if we, say, have an empty `storage` entry, we will return `StorageDriver: ""` on an Inspect request

the problem is, that is a lie. c/storage populates defaults for us, so we actually use overlay in that scenerio.

we don't encounter this normally, because if the toml field isn't specified (as it isn't by default), then we inherit the information from the default config
However, if a user (or an ignition config) overrides those fields to empty, we will listen to it as the source of truth, and never update to what is actually being used

Fix this by inheriting the values from c/storage on the validation step.

Signed-off-by: Peter Hunt <pehunt@redhat.com>