ttrpc: force pkg dependencies to older ones compatible with gogo/protobuf
The continerd/ttrpc package, used to support the runtimes of type "vm",
relies on the gogo/protobuf protocol buffer implementation.
Recently, google updated the golang protocol buffer implementation with
a newer and backward incompatible API:
https://blog.golang.org/protobuf-apiv2

The gogo/protobuf protocol has not been updated and is now incompatible
with the newer golang protocol buffer implementation.
Unfortunately, the containerd/ttrpc package imports also the package:
google.golang.org/genproto/googleapis/rpc/status
which is generated by means of the golang protocol buffer. With the
newer version, the generated code is no more compatible with
gogo/protobuf, leading to panic during execution of the ttrpc code.

While the ttrpc package needs a proper fix, this patch brings in a fork
of the containerd/ttrpc package depending upon an older version of the
googleapis/rpc/status package (compiled with the older protobuf API and
so compatible with the gogo/protobuf implementation) which has been forked
in a custom repo too. In this way, ttrpc code will import code that is
all compatible with the older protobuf specification.

Signed-off-by: Francesco Giudici <fgiudici@redhat.com>