Force pkg dependencies to older ones compatible with gogo/protobuf
The continerd/ttrpc package, used to support the runtimes of type "vm",
relies on the gogo/protobuf protocol buffer implementation.
Recently, google updated the golang protocol buffer implementation with
a newer and backward incompatible API:
https://blog.golang.org/protobuf-apiv2

The gogo/protobuf protocol has not been updated and is now incompatible
with the newer golang protocol buffer implementation.
Unfortunately, the containerd/ttrpc package imports also the package:
google.golang.org/genproto/googleapis/rpc/status
which is generated by means of the golang protocol buffer. With the
newer version, the generated code is no more compatible with
gogo/protobuf, leading to panic during execution of the ttrpc code.

While the ttrpc package needs a proper fix, this patch downgrades the
version of the golang/protobuf, genproto, and grpc.

This PR is a "backport" of:
https://github.com/cri-o/cri-o/commit/6bda9b5e079bdee399262330259970f1a40c16be

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Signed-off-by: Francesco Giudici <fgiudici@redhat.com>
Signed-off-by: Mrunal Patel <mrunalp@gmail.com>