Provide support for checkpoint and restore
This implements checkpoint and restore in cri-o.

Everything is based on the Podman implementation and copied from
Podman wherever possible (with small changes to fit into cri-o).

Currently the functionality is implemented as defined in the "Forensic
Container Checkpointing" KEP. Only container can be checkpointed and not
pods. The container will always continue to run after the checkpoint.

Using checkpoint via crictl could look something like this:

  # crictl checkpoint --export=archive.tar CTR_ID

The checkpointed container can be restored in a new pod using:

  # crictl create
  # crictl start CTR_ID

Signed-off-by: Adrian Reber <areber@redhat.com>