Makefile: fix vendor target
Commit 25eb7d841 added this target. Alas, the `go mod tidy` never
worked, as it was written as part of `export` statement:

	export GO111MODULE=on \
		$(GO) mod tidy && \
		...

which is the same as

	export GO111MODULE=on $(GO) mod tidy && ...

which exports a bunch of variables, such as `go`, `mod`, and `tidy`
(or would export them if they existed), but does not run `go mod tidy`.

The fix would be to add a semicolon after the `export` statement,
but it looks like GO111MODULE=on is not needed, so just remove it.

While at it, && are no longer needed either, so remove them as well.

For GOSUMDB, export it for the target via Makefile

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>