ctr.bats: fix jq checks
Commits a2ec1d40d64b and 247d465e8123e5 added a few checks
using jq in the following test cases:

 * ctr lifecycle
 * ctr execsync should not overwrite initial spec args
 * privileged ctr -- check for rw mounts

Alas, those checks do not work (and never worked); jq always succeeds.

This happened because

1. in `run cmd1 ... | cmd2` the part starting with the pipe
   character is not part of `run` statement;

2. `run` eats `cmd1 ...` output (into `$output` variable);

so `cmd2` is provided with empty input.

Now,

3. `jq` with empty input does not run any filters and thus succeeds
   (even with `-e`, see [1]).

The fix is to add a separate check that the output is not empty.

While at it, remove `run` where it's not needed from the other places
in those three tests we fix.

[1] https://github.com/stedolan/jq/issues/1628

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>