test/ctr: fix "privileged ctr -- check for rw mounts"
This [partially] fixes the test case.

It was broken from the very beginning, since it was supposed to run a
privileged container and check that /sys/fs/cgroup is mounted rw.

Commit d3a50d86f8803e21b ("privileged: set mounts to rw") added
test/testdata/container_config_privileged.json file, but it is not
used by the test (or any other tests). It also checks for ro mount
while it should check for rw (based on what I understood). Finally,
it used incorrect pod config while creating a container (not sure
if it makes any difference, just observing).

The test case was later amended to not check ro/rw bits since it was
failing on some configurations (see commit 78badc81ca4490).

This commit fixes the config used for container (and adds `edit_json`
function to which hopefully other tests will be eventually converted),
fixes the check for rw mount, and fixes the pod config used to create
container.

While at it, remove the never used container_config_privileged.json.

The issues that are not fixed (will be addressed separately) are:
 - non-working check via jq;
 - excessive use of `run`.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>