test: tidy image prefetch
There is a lot of repeated code with lots of options which is hard to
read. Introduce and use a few helper functions to clean up the mess,
as well as the (sorted) list of images we need for tests. In particular,
this fixes not copying fedora-ping-image to $TESTDIR.

The other thing is, function "pull_test_containers" is supposed to check
if images are available (and they are, since they are prefetched using
copyimg), pull them if they are not, and set some variables.

Some fixes to pull_test_containers are:
 - use the IMAGES list;
 - minimize the amount of `crictl inspecti` calls
   (as they generate lots of debug logs);
 - rename it to `check_images` (as it's not doing pull anymore);
 - `export` is not needed;
 - remove unused `*_IMAGEID` variables;

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>