test: rm edit_json, use jq directly
The `edit_json` introduced by commit 19e52142292ef2 was taken from
github.com/opencontainers/runc, where it is used to modify the json
file in place. Because of that, a contraption (implemented via awk)
is required to rewrite the file in place.

Here in cri-o we never modify files in place, so awk hack is not needed,
and without it, `edit_json` becomes `jq`, so there is no need to have
such a shallow wrapper.

While at it, remove | from the assignment, as we only have one such
element in the json.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>