runtime_vm: StopContainers() should not fail when the VM is shutdown
In the goroutine used to monitor whether the container was terminated or
not, we should not fail in case the VM was shutdown, as this is expected
to happen and will cause a ttrpc.ErrClosed.

runtime's wait() function, however, will returns errdefs.ErrNotFound
when a ttrpc.ErrCloses happens in order to avoid returning "ttrpc:
closed: unknown" (see previous commit) and that's the reason we just
check for errdefs.ErrNotFound and do not error out in that case.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>