runtime_vm: Ignore ttrpc.ErrClosed when removing a container
If the VM is down when removing the container, ttrpc.ErrClosed would be
returned and we'd return this error up in the chain. However, if the VM
is down, so is the container and we could simply ignore the error
reported, as already done in a few other parts of our code.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>