test: simplify check_journald check
The check_journald() was relying on presence of some packages
(needed to compile conmon with journald support), and outputtting
either 0 to 1 to stdout to indicate whether journald is supported.

Instead, let's do a runtime conmon check for journald support
(using some fake arguments so it won't error out if journald is
supported), and use the exit code check in its user.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>