test/network_ping: merge the two cases
Test setup and cleanup is rather heavy for crio, so having many small
test cases hurts performance.

This is a test-drive of an idea to merge small test cases, mostly to
see what the savings are, and as an RFC.

On my laptop I have the following results for

	time bats -t test/network_ping.bats

Before:

	1..2
	ok 1 Ping pod from the host
	ok 2 Ping pod from another pod

	real	0m37.787s
	user	0m11.697s
	sys	0m2.999s

After:

	1..1
	ok 1 Ping pod from the host / another pod

	real	0m28.048s
	user	0m6.207s
	sys	0m1.767s

So, every test case setup/cleanup takes about 10s wall clock time.

We have 178 tests now -- we can potentially easily save a few minutes.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>