oci: add Devices to allowed annotations structure
Before there was no way for a user to specify a device to be added to their containers.
The standard mechanism in k8s were clunky and didn't really work.

Instead, in the same way we allow userns to be specified with an annotation, add one such annotation for Devices.
Admins can restrict different runtime classes from this behavior with runtime classes.

also refactor devices tests to be all together in a separate file

Signed-off-by: Peter Hunt <pehunt@redhat.com>