test: disable crictl pull on create
We already preload a bunch of images before running tests, and it makes
sense to disable pull during crictl create, as it should save some time
and result in less flakiness in the tests (related to network glitches).

crictl 1.19 added this option, and "no pull on create" is the new
default. This reverts the hunk of commit 40b9d971fe which changes
the default.

For more details, see https://github.com/kubernetes-sigs/cri-tools/pull/627

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>