server: add userns mappings for spoofed infra
This is needed in case we spoof the infra container and have userns
enabled. In this case (*Server).getSandboxIDMappings() tries to read
infra container's /proc/$PID/uid_maps (and gid_maps) and it fails as
infra container PID is 0 in case it's spoofed.

But, if infra container's IDMappings are set, they are used instead of
reading from /proc.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>