runtime_vm: Ensure closeIOChan is not nil inside CloseStdin's function
This patch fixes the situation of `kubectl cp` hanging forever.

Debugging this issue I've faced a quite interesting situatuon where
after `closeIOChan` is closed, the `CloseStdin` functions "resumes"
while `closeIOChan` is **nil**, making the function to sit there and
wait forever.

Honestly, this bugs my mind quite a bit as I was under the impression
that all the everyone waiting for the channel to be closed would be
notified, would consume its content, and only after that it'd be
nullified.  Well, Today I learnt.

The simplest and more future proof way to work this issue around is to
ensure `closeIOChan` is not **nil** inside CloseStdin's function.

Thanks to @liubin, @haircommander, and @r4f4 for being available to
discuss the issue.

Fixes: #4353

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>