fork hosport kubernetes code
kubernetes upstream started the deprecation of dockershim, the
hostport manager code is under that folder, and is not clear
right now the direction that upstream is going to take regarding
this code.

Right now, t has an open bug and is not likely to be fixed, so it is better
for CRIO to use it and maintain it directly, until upstream takes a
final decision.

Signed-off-by: Antonio Ojea <aojea@redhat.com>