Make CNI setup instructions a bit more clear and fix nits
This calls out explicitly how to install the example CNI configurations (and also actually call out that the user has to do something here).

In addition to that this adds a few missing full stops and slightly alters the wording around the CNI version used, as the tutorial is explicitly **not** using the version at master.

Signed-off-by: Markus Thömmes <markusthoemmes@me.com>