ci: move golangci-lint from circleci to GH actions
It should be faster, and the side benefit of using github
actions is linter warnings will be shown in the diff
as annotations (similar to review comments).

To make it work without installing additional packages,
fix the build tags in .golangci.yml. In particular,
disable seccomp and device mapper, as they need some C
headers to be compiled.

Note that this is merely to be able to compile packages
used by crio, and since we are only linting crio itself,
and not the packages it imports, changing the tags not
used by crio doesn't matter much (if at all).

Note we still have `lint` target in Makefile, to be used
by developers (but not for CI).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>