github/PR template: add /kind ci, other
There was no category to characterize a PR or issue related to CI
(circle-ci setup and scripts, ansible jobs used by Jenkins, etc).
Let's add one.

There are some PRs that can't be clearly characterized as belonging
to a particular category (for example, this one). Add `/kind other`
for such.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>