circleci: fix cri-tools install
scripts/circle-setup currently fails like this while trying to install
cri-tools:

> sudo make BINDIR=/usr/bin install
> make: go: Command not found

This is caused by https://github.com/kubernetes-sigs/cri-tools/pull/684.
Basically, it now tries to build binaries under sudo and it fails
because go binary is not in the $PATH.

To fix, use `sudo -E "PATH=$PATH"`.

While at it, drop `make binaries` as this is now redundant.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>