contrib/test/int/e2e-features: skip Serial tests
This is mostly done to fix the flake caused by the test case

> Kubernetes e2e suite: [sig-network] IngressClass [Feature:Ingress] should not set default value if no default IngressClass [Serial]

which fails like this:

> ingressclasses.networking.k8s.io "ingressclass1" already exists

This test depends on a system state and thus can't be run in parallel.
It was marked as [Serial] recently (see [1]) but I just found out
we do not skip Serial tests in e2e-features.

Fix this.

[1] https://github.com/kubernetes/kubernetes/pull/93427

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>