Make NamespaceOption an internal type in sandbox
The fact that we re-use the protobuf devotions in the lower level
implementation details of CRI-O increases the burden to support multiple
Container Runtime Interface versions. This patch is one of a series of
patches to remove the protobuf type usage from the internal lib. This
allows us to create a middle layer later on to support two CRI versions
in parallel.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>