managed namespaces: move to dedicated package
The code for creating managed namespaces was clunky and all over the place.
For instance, we'd pass the server config to the sandbox and do the pinning there.
That confuses the design and creates really large function signatures.

Instead, create a dedicated package: nsmgr, that's responsible for choosing how namespaces work.

In addition:
- adapt test code to account for this change (drop sandbox mocks)
- drop unneeded methods
	- Get() was only used to check if the namespace was created. Instead, we should set a sandbox namespace if it's created
	- Close() was only used internally
	- Initialize{,d}() were also only used internally and not really needed

- rearrange file structure more logically
- rename types
	- move Namespace to internal, as no external libraries need access to it
	- change NamespaceIface to Namespace, as there's no need for us to specify the fact it's an interface

Signed-off-by: Peter Hunt <pehunt@redhat.com>