set inactive-or-failed CollectMode if appropriate
This commit introduces a bare bones internal/config/node package to support this feature.
Not cleaning failed scopes can cause mounts to fail, this piece of functionality is definitely a bug.
The rest of the internal/config/node package is not fixing a bug, so it's excluded.

currently, systemd leaks some scopes if they're failed.
This can be changed with the CollectMode annotation (which runc will set).
However, only systemd version >= 236 has support for this, so we need a new entry in internal/config/node to check whether systemd is the correct version

Signed-off-by: Peter Hunt <pehunt@redhat.com>