server: support setting raw unified cgroupv2 settings
allow to override the cgroup v2 unified configuration through a custom
annotation.  Multiple values are separed with `;`.

Each value can be specified in the base64 format.  If it is not a
valid base64 encoded string, then its original value is used.

The container is specified as part of the annotation name.

apiVersion: v1
kind: Pod
metadata:
  name: unified-test
  annotations:
    io.kubernetes.cri-o.UnifiedCgroup.cgroup-test: "memory.high=8000000;memory.low=100000"
spec:
  containers:
  - name: cgroup-test
    image: docker.io/alpine
    command: ["top"]

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>