cgroup: implement fix for swap memcg on cgroup v2
CRI-O assumed that the swap memcg is always present on cgroup v2 but
that is not always true as a kernel can be compiled without
CONFIG_SWAP or CONFIG_MEMCG_SWAP.

Implement the swap memcg check by looking into the current cgroup if
the memory.swap.current is present.  Now the assumption is that the
"memory" controller is enabled in the current cgroup.

Closes: https://github.com/containers/crun/issues/570

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>