Make rootless namespace unit test execution work
We should target to get the unit tests back to a state where we can
execute them in a rootless way. Right now the system user would not have
a chance to access pid 1's /proc/1/ns directory which make the test
fail. This patch uses the current pid, which should be accessible by the
current user executing the tests.

Can be verified by:
```
> make testunit TESTFLAGS=--focus=SandboxManagedNamespaces
```

Signed-off-by: Sascha Grunert <mail@saschagrunert.de>